ml/model/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.1",
    "optuna>=4.2.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "supabase>=2.14.0",
]
